// frontend/src/types/database.types.ts
export interface CaseDB {
  id: string;
  date_entered: Date | string | null;
  date_modified: Date | string | null;
  created_by: string | null;
  modified_user_id: string | null;
  assigned_user_id: string | null;
  deleted: number | null;
  case_number: number;
  name: string | null;
  account_id: string | null;
  contact_id: string | null;
  status: string | null;
  priority: string | null;
  type: string | null;
  description: string | null;
  resolution: string | null;
}